modified:   BackEnd/permovdb/src/main/java/com/permovdb/permovdb/PermovdbApplication.java 	modified:   BackEnd/permovdb/src/main/java/com/permovdb/permovdb/controller/MovieController.java 	modified:   BackEnd/permovdb/src/main/java/com/permovdb/permovdb/controller/UserController.java 	new file:   BackEnd/permovdb/src/main/java/com/permovdb/permovdb/domain/Cast.java 	new file:   BackEnd/permovdb/src/main/java/com/permovdb/permovdb/domain/CastMember.java 	new file:   BackEnd/permovdb/src/main/java/com/permovdb/permovdb/domain/DTO/MovieVideoDTO.java 	renamed:    BackEnd/permovdb/src/main/java/com/permovdb/permovdb/domain/UserDTO.java -> BackEnd/permovdb/src/main/java/com/permovdb/permovdb/domain/DTO/UserDTO.java 	modified:   BackEnd/permovdb/src/main/java/com/permovdb/permovdb/domain/Movie.java 	new file:   BackEnd/permovdb/src/main/java/com/permovdb/permovdb/domain/POJO/VideoMetadata.java 	modified:   BackEnd/permovdb/src/main/java/com/permovdb/permovdb/domain/User.java 	modified:   BackEnd/permovdb/src/main/java/com/permovdb/permovdb/domain/UserList.java 	new file:   BackEnd/permovdb/src/main/java/com/permovdb/permovdb/repository/CastMemberRepository.java 	modified:   BackEnd/permovdb/src/main/java/com/permovdb/permovdb/repository/MovieRepository.java 	new file:   BackEnd/permovdb/src/main/java/com/permovdb/permovdb/service/CastMemberService.java 	modified:   BackEnd/permovdb/src/main/java/com/permovdb/permovdb/service/MovieService.java 	new file:   BackEnd/permovdb/src/main/java/com/permovdb/permovdb/startup/Startup.java 	modified:   FrontEnd/perMovDb/package-lock.json 	modified:   FrontEnd/perMovDb/package.json 	modified:   FrontEnd/perMovDb/src/App.jsx 	modified:   FrontEnd/perMovDb/src/components/Card.jsx 	modified:   FrontEnd/perMovDb/src/components/CardPlate.jsx 	deleted:    FrontEnd/perMovDb/src/components/DetailsPage.jsx 	modified:   FrontEnd/perMovDb/src/components/Top.jsx 	modified:   FrontEnd/perMovDb/src/context/UserContext.jsx 	new file:   FrontEnd/perMovDb/src/pages/DetailsPage.jsx 	renamed:    FrontEnd/perMovDb/src/components/DiscoverPage.jsx -> FrontEnd/perMovDb/src/pages/DiscoverPage.jsx 	renamed:    FrontEnd/perMovDb/src/components/ErrorPage.jsx -> FrontEnd/perMovDb/src/pages/ErrorPage.jsx 	renamed:    FrontEnd/perMovDb/src/components/NewReleases.jsx -> FrontEnd/perMovDb/src/pages/NewReleases.jsx 	renamed:    FrontEnd/perMovDb/src/components/ProfilePage.jsx -> FrontEnd/perMovDb/src/pages/ProfilePage.jsx 	renamed:    FrontEnd/perMovDb/src/components/Upload.jsx -> FrontEnd/perMovDb/src/pages/Upload.jsx

+++ BackEnd/permovdb/src/main/java/com/permovdb/permovdb/domain/UserList.java
@@ -22,30 +22,30 @@ public class UserList {
     @ElementCollection
     @CollectionTable(name = "user_movie_watchlist", joinColumns = @JoinColumn(name = "user_id"))
     @Column(name = "movie_id")
-    private Set<Long> movieWatchlist = new HashSet<>();
+    private Set<Integer> movieWatchlist = new HashSet<>();
 
     @ElementCollection
     @CollectionTable(name = "user_movie_watchedlist", joinColumns = @JoinColumn(name = "user_id"))
     @Column(name = "movie_id")
-    private Set<Long> movieWatchedlist = new HashSet<>();
+    private Set<Integer> movieWatchedlist = new HashSet<>();
 
     @ElementCollection
     @CollectionTable(name = "user_movie_lovedlist", joinColumns = @JoinColumn(name = "user_id"))
     @Column(name = "movie_id")
-    private Set<Long> movieLovedlist = new HashSet<>();
+    private Set<Integer> movieLovedlist = new HashSet<>();
 
     @ElementCollection
     @CollectionTable(name = "user_tv_watchlist", joinColumns = @JoinColumn(name = "user_id"))
     @Column(name = "tv_id")
-    private Set<Long> tvWatchlist = new HashSet<>();
+    private Set<Integer> tvWatchlist = new HashSet<>();
 
     @ElementCollection
     @CollectionTable(name = "user_tv_watchedlist", joinColumns = @JoinColumn(name = "user_id"))
     @Column(name = "tv_id")
-    private Set<Long> tvWatchedlist = new HashSet<>();
+    private Set<Integer> tvWatchedlist = new HashSet<>();
 
     @ElementCollection
     @CollectionTable(name = "user_tv_lovedlist", joinColumns = @JoinColumn(name = "user_id"))
     @Column(name = "tv_id")
-    private Set<Long> tvLovedlist = new HashSet<>();
+    private Set<Integer> tvLovedlist = new HashSet<>();
 }
